+++ app/src/main/java/org/simple/clinic/settings/SettingsScreen.kt
@@ -35,7 +35,7 @@ class SettingsScreen : BaseScreen<
     SettingsModel,
     SettingsEvent,
     SettingsEffect,
-    Unit>(), SettingsUi, UiActions {
+    SettingsViewEffect>(), SettingsUi, UiActions {
 
   @Inject
   lateinit var router: Router
@@ -79,7 +79,12 @@ class SettingsScreen : BaseScreen<
 
   override fun createInit() = SettingsInit()
 
-  override fun createEffectHandler(viewEffectsConsumer: Consumer<Unit>) = settingsEffectHandler.create(this).build()
+  override fun createEffectHandler(viewEffectsConsumer: Consumer<SettingsViewEffect>) =
+      settingsEffectHandler
+          .create(
+              viewEffectsConsumer = viewEffectsConsumer
+          )
+          .build()
 
   override fun createUpdate() = SettingsUpdate()
 
@@ -88,6 +93,8 @@ class SettingsScreen : BaseScreen<
       .cast<SettingsEvent>()
 
   override fun uiRenderer() = SettingsUiRenderer(this)
+
+  override fun viewEffectHandler() = SettingsViewEffectHandler(this)
 
   override fun onAttach(context: Context) {
     super.onAttach(context)