+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - Bump Flipper to v0.131.0
 - Add benchmark test for creating a new bp entry 
 - Migrate `CustomDrugEntrySheet` to user view effects
+- Migrate `SettingsScreen` to use view effects
 
 ### Changes